refactor: just check for a group of two
We don't need to check so many criterias for the groups.
The conditions exclude all numbers that doesn't has a group
of two. So instead of checking for numbers with groups of
digits and then imposing rules on them, we can check for
numbers with groupings of two digits because those are the
only numbers that are allowed anyway.